+++ blocklets/ai-studio/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.2.15 (2024-5-28)
+
+- fix: change dialog to popper
+
+## 0.2.14 (2024-5-27)
+
+- chore: return project primary color for agents
+
 ## 0.2.13 (2024-5-27)
 
 - chore: support choose a blocklet react component to render output appearance (#1016)